+++ spec/cryptozoologist/generator_spec.rb
@@ -107,11 +107,27 @@ describe Cryptozoologist do
 
   context '#name' do
     before do
-      @person = Cryptozoologist::Dictionaries::People::FirstName.list.sample
+      @person = Cryptozoologist.name
     end
 
-    it 'should return a string' do
+    it 'should return datatype string' do
       expect(@person).to be_instance_of(String)
+    end
+    
+    it 'should contain an animal' do 
+      has_animal = false
+      Cryptozoologist::Dictionary.animals.each do |animal|
+        has_animal = true if @person.downcase.include?(animal)
+      end
+      expect(has_animal).to be true 
+    end
+    
+    it 'should contain an adjective for the animal' do 
+      has_adjective = false
+      Cryptozoologist::Dictionaries::People::LastName.list.each do |adjective|
+        has_adjective = true if @person.include?(adjective)
+      end
+      expect(has_adjective).to be true 
     end
   end
 end